Create new components for steps inputs
