+++ public/controllers/register-agent/components/group-input/group-input.scss
