+++ public/controllers/register-agent/components/group-input/group-input.tsx
