+++ public/controllers/register-agent/components/os-selector/checkbox-group/checkbox-group.scss
