+++ public/controllers/register-agent/components/os-selector/os-card/os-card.test.tsx
