+++ public/controllers/register-agent/containers/register-agent/register-agent.scss
@@ -1,6 +1,6 @@
 .container {
   box-sizing: border-box;
-  max-height: 1271px;
+  min-height: 1271px;
   margin-top: 44px;
   background: #ffffff;
   border: 1px solid rgba(52, 55, 65, 0.2);
